feat: update GCP docker build and push conditions (#111)
* feat: Add conditional check for open PR in GCP Docker workflow

The code changes in this commit add a conditional check for an open pull request (PR) in the GCP Docker workflow. If the push event is for an open PR, the workflow will skip the current run. This check is implemented by retrieving the PR data using the GitHub API and checking the state of the PR.

* chore: Update generated build push workflow name

* feat(gcp-docker): Simplify build workflow logic for push events

* chore: Add 'needs' dependency for monolith job in GCP Docker workflow

* chore: Add 'monolith' as a dependency for the GCP Docker workflow