Refactor sync-vscode-extensions.yml workflow
This commit refactors the sync-vscode-extensions.yml workflow by updating the version of the 'create-pull-request' action from v6 to v7. It also adds a new step to the workflow to disable persisting credentials during the checkout process. These changes aim to improve the synchronization of VSCode extensions between the .devcontainer/devcontainer.json and .vscode/extensions.json files for the current branch.